+++ roles/collections_server.js
@@ -1,0 +1,7 @@
+(function() {
+  'use strict';
+
+  // Create default indexes for authItems collection
+  Meteor.authItems._ensureIndex('name', { unique: 1 })
+
+}());