+++ roles/subscriptions.js
@@ -1,0 +1,10 @@
+(function() {
+  'use strict';
+
+  /**
+   * Subscribe to auth items.
+   */
+  Meteor.subscribe('authItems');
+
+}());
+