+++ roles/tests/client.js
@@ -1,18 +1,17 @@
-;(function () {
-
-  "use strict";
+(function () {
+  'use strict';
 
   // Mock Meteor.user() for role testing
   Meteor.user = function () {
     return {
       _id: 'testId',
-      roles: ['user','manage-users']
-    }
-  }
+      authItems: ['user', 'manage-users']
+    };
+  };
 
 
   Tinytest.add(
-    'roles - can check current users roles via template helper', 
+    'roles - can check current users roles via template helper',
     function (test) {
       var isInRole = Roles._handlebarsHelpers.isInRole,
           expected,
@@ -23,11 +22,10 @@
       expected = true
       actual = isInRole('admin, manage-users')
       test.equal(actual, expected)
-      
+
       expected = false
       actual = isInRole('admin')
       test.equal(actual, expected)
     })
-
 
 }());