+++ History.md
@@ -1,3 +1,7 @@
 ## v1.0.0
 
-* To be released!
+* Add support for permissions
+* Allow permissions and roles to be nested in other roles
+* Rewrite core app to deal with Auth Items, instead of just roles
+* Replaced 'Roles' global with a 'AuthManager' global
+* NOTE: package forked from https://github.com/alanning/meteor-roles by Adrian Lanning